Synced Streamlit App with OpenAI API, Removed irrelevant Flask files,
