--- static/css/dashboard_styles.css
@@ -1,102 +1,0 @@
-/* Import the simpler font */
-@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;500&display=swap');
-
-/* Reset default browser styles */
-body, h1, p, a {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-/* Apply background styles with a dark purple color */
-body {
-    font-family: 'Roboto', sans-serif;
-    color: #f0f0f0;
-    background-color: #d8cbe5; /* light-purple */
-}
-
-/* Header styling */
-.dashboard-header {
-    text-align: center;
-    padding: 50px 0;
-}
-
-/* Style the main heading */
-h1 {
-    font-size: 48px;
-    color: #5c538f; /* Pastel purple */
-    margin-bottom: 10px;
-}
-
-/* Highlight the 'AI' part of the heading */
-.highlight {
-    color: #9575cd; /* Medium pastel purple */
-}
-
-/* Welcome message styling */
-.welcome {
-    font-size: 16px;
-    color: #2f023c; /* Light pastel purple */
-    margin-bottom: 40px;
-}
-
-/* Content section styling */
-.content {
-    width: 90%;
-    max-width: 600px;
-    margin: 0 auto;
-    text-align: center;
-    background-color: #5c538f; /* Slightly lighter dark purple */
-    padding: 40px;
-    border-radius: 10px;
-    box-shadow: 0 0 15px rgba(0, 0, 0, 0.5);
-}
-
-/* Placeholder info text styling */
-.info {
-    font-size: 18px;
-    color: #d1c4e9; /* Lighter pastel purple */
-    margin-bottom: 30px;
-}
-
-/* Logout button styling */
-.logout-btn {
-    display: inline-block;
-    padding: 10px 20px;
-    background-color: #9575cd; /* Medium pastel purple */
-    border: none;
-    border-radius: 4px;
-    color: #ffffff;
-    font-size: 18px;
-    text-decoration: none;
-    cursor: pointer;
-    transition: background-color 0.3s;
-}
-
-/* Change background color on hover for logout button */
-.logout-btn:hover {
-    background-color: #7e57c2; /* Slightly darker purple */
-}
-
-/* Footer styling */
-.footer {
-    text-align: center;
-    margin-top: 50px;
-    color: #2f023c;
-    font-size: 14px;
-}
-
-/* Responsive design for smaller screens */
-@media (max-width: 480px) {
-    h1 {
-        font-size: 36px;
-    }
-    .content {
-        margin: 20px;
-        padding: 20px;
-    }
-    .logout-btn {
-        font-size: 16px;
-        padding: 8px 16px;
-    }
-}
