--- static/css/passcode_styles.css
@@ -1,138 +1,0 @@
-/* Import the Roboto font */
-@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;500&display=swap');
-
-/* Reset default browser styles */
-body, h1, h2, p, form, input {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-/* Apply background styles with a light purple color */
-body {
-    font-family: 'Roboto', sans-serif;
-    color: #2f023c;
-    background-color: #d8cbe5; /* Light purple */
-    text-align: center;
-    padding: 20px;
-}
-
-/* Style the main heading */
-.title {
-    font-size: 64px;
-    color: #5c538f; /* Dark purple */
-    margin-bottom: 20px;
-}
-
-/* Highlight the 'AI' part of the heading */
-.highlight {
-    color: #9575cd; /* Medium pastel purple */
-}
-
-/* Style the description */
-.description {
-    font-size: 16px;
-    margin: 0 auto 40px auto;
-    max-width: 800px;
-    line-height: 1.5;
-    color: #2f023c; /* Dark text color for readability */
-    padding: 0 20px;
-}
-
-/* Center the container and style it */
-.container {
-    width: 90%;
-    max-width: 400px;
-    margin: 0 auto;
-    background-color: #5c538f; /* Dark purple */
-    padding: 40px;
-    border-radius: 10px;
-    text-align: center;
-    box-shadow: 0 0 15px rgba(0, 0, 0, 0.2);
-}
-
-/* Style the instructions */
-.instructions {
-    font-size: 20px;
-    margin-bottom: 20px;
-    color: #d1c4e9; /* Lighter pastel purple */
-}
-
-/* Style the error message */
-.error {
-    color: #ff5252; /* Light red for error messages */
-    margin-bottom: 20px;
-    font-weight: bold;
-}
-
-/* Style the form */
-form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-/* Style the input fields */
-input[type="password"] {
-    width: 100%;
-    padding: 12px 15px;
-    margin-bottom: 20px;
-    border: 1px solid #616161; /* Gray border */
-    border-radius: 4px;
-    font-size: 16px;
-    background-color: #424242; /* Dark gray */
-    color: #f0f0f0;
-    transition: border-color 0.3s;
-}
-
-/* Change border color on focus */
-input[type="password"]:focus {
-    border-color: #9575cd; /* Medium pastel purple */
-    outline: none;
-}
-
-/* Style the submit button */
-input[type="submit"] {
-    width: 100%;
-    padding: 12px 15px;
-    background-color: #9575cd; /* Medium pastel purple */
-    border: none;
-    border-radius: 4px;
-    color: #ffffff;
-    font-size: 18px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-}
-
-/* Change background color on hover */
-input[type="submit"]:hover {
-    background-color: #7e57c2; /* Slightly darker purple */
-}
-
-/* Style the footer */
-.footer {
-    text-align: center;
-    margin-top: 50px;
-    color: #2f023c;
-    font-size: 14px;
-}
-
-/* Responsive design for smaller screens */
-@media (max-width: 480px) {
-    .title {
-        font-size: 36px;
-    }
-    .description {
-        font-size: 16px;
-        padding: 0 10px;
-    }
-    .container {
-        padding: 20px;
-    }
-    .instructions {
-        font-size: 18px;
-    }
-    input[type="password"], input[type="submit"] {
-        font-size: 14px;
-    }
-}
