--- templates/dashboard.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>PensieveAI Dashboard</title>
-    <meta charset="UTF-8">
-    <!-- Link to Google Fonts for Roboto font -->
-    <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500&display=swap" rel="stylesheet">
-    <!-- Link to external CSS file -->
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/dashboard_styles.css') }}">
-</head>
-<body>
-    <div class="dashboard-header">
-        <h1>Pensieve<span class="highlight">AI</span> Dashboard</h1>
-        <p class="welcome">Welcome to the main application interface, where you can begin uploading files for analysis.</p>
-    </div>
-    
-    <div class="content">
-        <!-- Placeholder for future file upload form and analysis features -->
-        <p class="info">⚡ File Upload and Analysis features coming soon! ⚡</p>
-        <a href="{{ url_for('logout') }}" class="logout-btn">Logout</a>
-    </div>
-
-    <div class="footer">
-        © 2024 PensieveAI. All rights reserved.
-    </div>
-</body>
-</html>
